pyproject.toml
[tool.poetry]
name = "pharmade"
version = "0.1.0"
description = ""
authors = ["faizeraza <faizeraza4us@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
streamlit = "^1.30.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
